requirements.txt
edge-tts
SpeechRecognition
PyAudio
send2trash
playsound3
psutil
wmi
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai
wikipedia
pygame
fuzzywuzzy
python-Levenshtein
selenium
webdriver-manager
requests
beautifulsoup4
pandas
keyboard
openpyxl